MOved methods to `Photo`, removed `'raw'`, fixed some exception errors.
